Moved files to better understand and start creating validations on route
