--- Exercise-2/.gitignore
@@ -1,1 +1,0 @@
-node_modules
