--- Exercise-2/Dockerfile
@@ -1,5 +1,0 @@
-FROM node:latest
-ADD . /app
-WORKDIR /app
-RUN npm install
-CMD ["node", "index.js"]
