--- Exercise-2/docker-compose.yml
@@ -1,11 +1,0 @@
-version: '3'
-services:
-  exercise1:
-    build: .
-    ports:
-     - "9001:9001"
-  messageapp:
-    image: "cabify/backend-bootcamp-messageapp:latest"
-    ports:
-      - "3000:3000"
-
