--- Exercise-2/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "exercise-1",
-  "version": "1.0.0",
-  "description": "first exercise of cabify bootcamp",
-  "main": "index.js",
-  "scripts": {
-    "start": "index.js"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
-    "express": "^4.16.3"
-  }
-}
