+++ sendMessage.js
@@ -1,5 +1,5 @@
 const axios = require('axios')
-const API_URL = 'http://exercise-1_messageapp_1:3000/message'
+const API_URL = 'http://teo_diaz_messageapp_1:3000/message'
 
 const sendMessage = (req) =>{
   const { destination, body } = req.body;
